# Speedup NestedSiblinFileCollection when possible