In my tests, this gives a speed-up of a factor of more than 60 for a nested sibling file collection of 4000 files.
```console
$ python test.py
NestedSiblingFileCollection(len=4000, threshold=1.0, collections=4)
True
Total: 62.249094 seconds

$ python test.py
NestedSiblingFileCollection(len=4000, threshold=1.0, collections=4)
True
Total: 0.858903 seconds
```

Testing script:
```python
import law
from law import NestedSiblingFileCollection
import time

law.contrib.load("wlcg")


targets = []
for j in range(4):
    for i in range(1000):
        i += j*1000
        targets.append(law.wlcg.WLCGFileTarget(f"HerwigRun/Dijets_LO/NPoff/{j}/Dijets_LOjob{i}.tar.bz2"))

collection = NestedSiblingFileCollection(targets)
print(collection)
start = time.time()
print(collection.complete())
end = time.time()
print(f"Total: {end-start:2f} seconds")
```
Testing law.cfg:
```
[target]
default_wlcg_fs = wlcg_fs

[wlcg_fs]
base = root://cmsxrootd-redirectors.gridka.de:1094//store/user/mhorzela/HerwigMC
```

I think this can be implemented way cleaner, but I did not come up with a good idea on how to get the correct keys otherwise. Maybe @riga knows a more general way for also achieving this speed-up, when the keys are required.