@cverstege 2024-11-14
You do need a valid CMS voms proxy to be able to read the files though.