@riga 2024-11-25
That looks good! To simplify things even further, could you just use

```python
        # reuse state iteration of wrapped collections
        for coll in self.collections:
            iter_kwargs = {
                "existing": existing,
                "optional_existing": optional_existing,
                "keys": keys,
                "unpack": unpack,
            }
            if isinstance(coll, SiblingFileCollectionBase) and coll in basenames:
                iter_kwargs["basenames"] = basenames[coll]
            for obj in coll._iter_state(**iter_kwargs):
                yield obj
```

and see if that works for you? fyi, the `exists` helper is also not needed any longer with this change.