@cverstege 2024-11-26
Looks good with my testing script. I will also test this in a production setup today.