@cverstege 2024-11-26
Running tasks using NestedSiblingFileCollections as inputs seems to be fine.

Timing just the `--print-status` on a dev version of https://github.com/HerrHorizontal/mc-run:

Old 106.94s:
```cmd
$ time law run RivetMerge --mc-generator herwig --campaign Dijets_LO --mc-setting NPoff --number-of-gen-jobs 40000 --rivet-analyses '["Dijet_3_highpt", "MC_XS", "MC_WEIGHTS"]' --print-status 1
WARNING: generation.framework.utils - KeyError: 'RIVET_ANALYSIS_PATH', variable undefined on local host!
print task status with max_depth 1 and target_depth 0

0 > RivetMerge(campaign=Dijets_LO, mc_setting=NPoff, rivet_analyses=["Dijet_3_highpt", "MC_XS", "MC_WEIGHTS"], chunk_size=100, mc_generator=herwig, skip_max_uncertainty=0.25, number_of_gen_jobs=40000)
│     WLCGFileTarget(fs=wlcg_fs, path=/RivetMerge/herwig/Dijets_LO/NPoff/Dijets_LO/Dijet_3_highpt_MC_WEIGHTS_MC_XS.yoda)
│       absent
│
├──1 > RunRivet(effective_workflow=htcondor, branch=-1, campaign=Dijets_LO, mc_setting=NPoff, files_per_job=10, rivet_analyses=["Dijet_3_highpt", "MC_XS", "MC_WEIGHTS"], mc_generator=herwig, number_of_gen_jobs=4
│      0000, workflow=htcondor)
│        jobs: LocalFileTarget(fs=local_fs, path=/home/cverstege/jobs/RunRivet__1__Dijets_LO_cb5f7a8f0b/htcondor_jobs_0To4000.json, optional)
│          existent
│        collection: NestedSiblingFileCollection(len=4000, threshold=4000.0, collections=4)
│          existent (4000/4000)
INFO: generation.tasks.RivetBuild - Built-in Rivet analysis MC_XS. Skip building...
INFO: generation.tasks.RivetBuild - Built-in Rivet analysis MC_WEIGHTS. Skip building...
│
└──1 > RivetBuild(effective_workflow=htcondor, branch=-1, rivet_analyses=["Dijet_3_highpt", "MC_XS", "MC_WEIGHTS"], workflow=htcondor)
         jobs: LocalFileTarget(fs=local_fs, path=/home/cverstege/jobs/RivetBuild__1__False_dbd5747934/htcondor_jobs_0To1.json, optional)
           absent
         collection: SiblingFileCollection(len=1, threshold=1.0, fs=wlcg_fs, dir=/RivetBuild/Alma9)
           existent (1/1)
law run RivetMerge --mc-generator herwig --campaign Dijets_LO --mc-setting     106.94s user 3.92s system 97% cpu 1:54.19 total
```

New 38.07s:
```
$ time law run RivetMerge --mc-generator herwig --campaign Dijets_LO --mc-setting NPoff --number-of-gen-jobs 40000 --rivet-analyses '["Dijet_3_highpt", "MC_XS", "MC_WEIGHTS"]' --print-status 1
WARNING: generation.framework.utils - KeyError: 'RIVET_ANALYSIS_PATH', variable undefined on local host!
print task status with max_depth 1 and target_depth 0

0 > RivetMerge(campaign=Dijets_LO, mc_setting=NPoff, rivet_analyses=["Dijet_3_highpt", "MC_XS", "MC_WEIGHTS"], chunk_size=100, mc_generator=herwig, skip_max_uncertainty=0.25, number_of_gen_jobs=40000)
│     WLCGFileTarget(fs=wlcg_fs, path=/RivetMerge/herwig/Dijets_LO/NPoff/Dijets_LO/Dijet_3_highpt_MC_WEIGHTS_MC_XS.yoda)
│       absent
│
├──1 > RunRivet(effective_workflow=htcondor, branch=-1, campaign=Dijets_LO, mc_setting=NPoff, files_per_job=10, rivet_analyses=["Dijet_3_highpt", "MC_XS", "MC_WEIGHTS"], mc_generator=herwig, number_of_gen_jobs=4
│      0000, workflow=htcondor)
│        jobs: LocalFileTarget(fs=local_fs, path=/home/cverstege/jobs/RunRivet__1__Dijets_LO_cb5f7a8f0b/htcondor_jobs_0To4000.json, optional)
│          existent
│        collection: NestedSiblingFileCollection(len=4000, threshold=4000.0, collections=4)
│          existent (4000/4000)
INFO: generation.tasks.RivetBuild - Built-in Rivet analysis MC_XS. Skip building...
INFO: generation.tasks.RivetBuild - Built-in Rivet analysis MC_WEIGHTS. Skip building...
│
└──1 > RivetBuild(effective_workflow=htcondor, branch=-1, rivet_analyses=["Dijet_3_highpt", "MC_XS", "MC_WEIGHTS"], workflow=htcondor)
         jobs: LocalFileTarget(fs=local_fs, path=/home/cverstege/jobs/RivetBuild__1__False_dbd5747934/htcondor_jobs_0To1.json, optional)
           absent
         collection: SiblingFileCollection(len=1, threshold=1.0, fs=wlcg_fs, dir=/RivetBuild/Alma9)
           existent (1/1)
law run RivetMerge --mc-generator herwig --campaign Dijets_LO --mc-setting     38.07s user 4.08s system 92% cpu 45.564 total
```